refactor: [M3-6320] - MUI v5 Migration - Components > DialogTitle
