--- packages/manager/src/components/DialogTitle/index.tsx
@@ -1,2 +1,0 @@
-import _DialogTitle from './DialogTitle';
-export default _DialogTitle;
